Converted MVC to MVVM
