+++ Person-Info/Model/PersonModel.swift
@@ -5,6 +5,7 @@
 //  Created by Tejasv Singh on 9/12/25.
 //
 import Foundation
+
 struct person: Decodable {
     let name: String
     let gender: String
